Fixed a bug where stage was being disposed twice
The Stage that had been initialised in GameMenu.java was being
passed to BuildingMenu.java was a parameter in the constructor.
When the BuildingMenu.dispose() method was called, the stage was
being disposed and then in GameMenu.dispose() the stage tried
to be disposed of again which caused an Exception.