Mostly fixed building placements
You now can't overlap buildings, and they're confined to the map.

However, I am not smart enough to think of a better way to check for collisions, as we add buildings to an array before doing anything, which makes it really difficult to use in built overlap/collision features. I couldn't summon the effort to sort all the boolean clauses by x values to prevent index out of range errors, so as a temporary fix I've just prevented placing buildings at the left.

I also prevented (intentionally for once) placing buildings at the top, as to not get in the way of any information we display up there.

Also the current terrible overlap check will only work for buildings with the same dimensions.