Added building counters
Added building counters, both as variables in gameModel, and as text on screen.

Added a new font ( the same one but smaller), as well as moved the font into gameModel (out of screenmanager).

Also figured out and fixed the bug caused by calling drawBuildingMenu, and fixed the menu being slightly too narrow only when not in fullscreen for some reason.