Altered construction timers
Altered the black font to be bold, moved the code placing the text into draw, rather than drawmapobjects, allowing the text to be drawn after the building is drawn, so it is visible.

Then added 2 arrays, one that stores the tile coords of the buildings under construction, and one that temporarily stores those that have finished construction. Then added a for loop that iterates through the first loop to draw the construction timers, and adds the finished ones to the found list, to be removed after, such that there is no ConcurrentModificationException.